Index

Index

Base.maximum — Method.

maximum(A::ClimGrid)

Compute the maximum value of ClimGrid A

source

Base.minimum — Method.

minimum(A::ClimGrid)

Compute the minimum value of ClimGrid A

source

Base.std — Method.

std(A::ClimGrid)

Compute the standard deviation of ClimGrid A

source

Base.var — Method.

var(A::ClimGrid)

Compute the variance of ClimGrid A

source

ClimateTools.annualmax — Method.

annualmax(data::AbstractArray, time::StepRange{Date,Base.Dates.Day})

Annual maximum of array data.

Let data[i,j] be daily time serie on day i in year j. Extract the highest value for year j.

source

ClimateTools.annualmax — Method.

annualmax(C::ClimGrid)

Annual maximum of array data.

Let data[i,j] be daily time serie on day i in year j. Extract the highest value for year j.

source

ClimateTools.annualmean — Method.

annualmean(data::AbstractArray, time::StepRange{Date,Base.Dates.Day})

Annual mean of array data.

Let data[i,j] be daily time serie on day i in year j. Calculate the mean value for year j.

source

ClimateTools.annualmean — Method.

annualmean(C::ClimGrid)

Annual mean of array data.

Let data[i,j] be daily time serie on day i in year j. Calculate the mean value for year j.

source

ClimateTools.annualmin — Method.

annualmin(data::AbstractArray, time::StepRange{Date,Base.Dates.Day})

Annual minimum of array data.

Let data[i,j] be daily time serie on day i in year j. Extract the lowest value for year j.

source

ClimateTools.annualmin — Method.

annualmin(C::ClimGrid)

Annual minimum of array data.

Let data[i,j] be daily time serie on day i in year j. Extract the lowest value for year j.

source

ClimateTools.annualsum — Method.

annualsum(data::AbstractArray, time::StepRange{Date,Base.Dates.Day})

Value of annual sum of array data.

Let data[i,j] be daily time serie on day i in year j. Sums daily values for year j.

source

ClimateTools.annualsum — Method.

annualsum(C::ClimGrid)

Annual sum of array data.

Let data[i,j] be daily time serie on day i in year j. Sums daily values for year j.

source

ClimateTools.applymask — Method.

applymask(A::AbstractArray{N, n}, mask::AbstractArray{N, n})

This function applies a mask on the array A. Return an AbstractArray{N, n}.

source

ClimateTools.buildtimevec — Method.

buildtimevec(str::String)Y

Construct the time vector from the netCDF file str

source

ClimateTools.customthresover — Method.

customthresover(data::AbstractArray, time::StepRange{Date,Base.Dates.Day}, thres)

customthresover, annual number of days over a specified threshold.

Let TS[i,j] be a daily time serie value on day i in year j. Count the number of days where:

TS[i,j] > thres.

source

ClimateTools.customthresover — Method.

customthresover(C::ClimGrid)

customthresover, annual number of days over a specified threshold.

Let TS[i,j] be a daily time serie value on day i in year j. Count the number of days where:

TS[i,j] > thres.

source

ClimateTools.customthresunder — Method.

customthresunder(data::AbstractArray, time::StepRange{Date,Base.Dates.Day}, thres)

customthresover, annual number of days under a specified threshold.

Let TS[i,j] be a daily time serie value on day i in year j. Count the number of days where:

TS[i,j] < thres.

source

ClimateTools.customthresunder — Method.

customthresunder(C::ClimGrid)

customthresover, annual number of days under a specified threshold.

Let TS[i,j] be a daily time serie value on day i in year j. Count the number of days where:

TS[i,j] < thres.

source

ClimateTools.daysabove10 — Method.

daysabove10(data::AbstractArray, timevector::StepRange{Date,Base.Dates.Day})

Annual number with temperature >= 10 Celsius. This function returns a boolean vector. true if the data is higher or equal to 1 and false otherwise.

source

ClimateTools.daysabove10 — Method.

daysabove10(C::ClimGrid)

Annual number of days with temperature >= 10 Celsius. This function returns a ClimGrid.

source

ClimateTools.frostdays — Method.

frostdays(data::AbstractArray, time::StepRange{Date,Base.Dates.Day})

FD, Number of frost days: Annual count of days when TN (daily minimum temperature) < 0 Celsius.

Let TN[i,j] be daily minimum temperature on day i in year j. Count the number of days where:

TN[i,j] < 0 Celsius.

source

ClimateTools.frostdays — Method.

frostdays(C::ClimGrid)

FD, Number of frost days: Annual count of days when TN (daily minimum temperature) < 0 Celsius.

Let TN[i,j] be daily minimum temperature on day i in year j. Count the number of days where:

TN[i,j] < 0 Celsius.

source

ClimateTools.icingdays — Method.

icingdays(data::AbstractArray, time::StepRange{Date,Base.Dates.Day})

ID, Number of summer days: Annual count of days when TX (daily maximum temperature) < 0 degree Celsius.

Let TX[i,j] be daily maximum temperature on day i in year j. Count the number of days where:

TX[i,j] < 0 Celsius.

source

ClimateTools.icingdays — Method.

icingdays(C::ClimGrid)

ID, Number of summer days: Annual count of days when TX (daily maximum temperature) < 0 degree Celsius.

Let TX[i,j] be daily maximum temperature on day i in year j. Count the number of days where:

TX[i,j] < 0 Celsius.

source

ClimateTools.inpoly — Method.

inpoly(p, poly::Matrix)

Determines if a point is inside a polygon.

p – point (x,y) or [x,y]

poly – polygon vertices [x1 x2 ... xn x1 y1 y2 ... yn y1]

Returns true if point has an odd winding number. This should label points as exterior which are inside outcrops. See test for a test.

Author: Github "Mauro3" / "Mauro"

source

ClimateTools.inpolygrid — Method.

inpolygrid(lon, lat, poly::AbstractArray{N,2} where N)

Used to test a grid of points. Returns a mask of ones and NaNs of the same size as lon and lat.

source

ClimateTools.interp_climgrid — Method.

C = interp_climgrid(A::ClimGrid, londest::AbstractArray{N, 1} where N, latdest::AbstractArray{N, 1} where N)A

This function interpolate ClimGrid A onto lat-lon grid defined by londest and latdest vector.

source

ClimateTools.interp_climgrid — Method.

C = interp_climgrid(A::ClimGrid, B::ClimGrid; method="linear", min=[], max=[])

This function interpolate ClimGrid A onto the lon-lat grid of ClimGrid B, where A and B are ClimGrid. Available methods for interpolation are "linear" (default), "nearest" and "cubic".

Min and max optional keyword are used to constraint the results of the interpolation. For example, interpolating bounded fields can lead to unrealilstic values, such as negative precipitation. In that case, one would use min=0.0 to convert negative precipitation to 0.0.

source

ClimateTools.mapclimgrid — Method.

mapclimgrid(C::ClimGrid; region::String="auto", poly, level, mask, caxis, start_date::Date, end_date::Date)

Maps the time-mean average of ClimGrid C.

Optional keyworkd includes precribed regions (see list below), spatial clipping by polygon (keyword poly) or mask (keyword mask, an array of NaNs and 1.0 of the same dimension as the data in ClimGrid C), start_date and end_date.

Arguments

Regions

Europe

NorthAmerica

Canada

Quebec, QuebecNSP

Americas

World, WorldAz, WorldEck4

Greenwich

source

ClimateTools.mapclimgrid — Method.

mapclimgrid(; region::String="auto", poly, level, mask, caxis, start_date::Date, end_date::Date)

Empty map generator, when called without a ClimGrid as the positional argument.

source

ClimateTools.meshgrid — Method.

X, Y = meshgrid{T}(vx::AbstractVector{T}, vy::AbstractVector{T})

This function creates a 2-D mesh-grid in a format consistent with Matlab's function meshgrid(). XV and YV are vectors.

source

ClimateTools.nc2julia — Method.

nc2julia(file::String, variable::String; poly = Array{Float64}([]), start_date::Date, end_date::Date, data_units::String = "")

Returns a ClimGrid type with the data in file of variable var inside the polygon poly. Metadata is built-in the ClimGrid type, from the netCDF attributes.

Inside the ClimgGrid type, the data is stored into an AxisArray data type, with time, longitude/x and latitude/y dimensions.

The polygon provided should be in the -180, +180 longitude format. If the polygon crosses the International Date Line, the polygon should be splitted in multiple parts (i.e. multi-polygons).

Options for data_units are for precipitation : "mm", which converts the usual "kg m-2 s-1" unit found in netCDF files. For temperature : "Celsius", which converts the usual "Kelvin" unit.

Note: nc2julia is based on CF conventions (http://cfconventions.org/). If you are unable to read the netCDF file with nc2julia, the user will need to read it with low-level functions available in the NetCDF.jl package (https://github.com/JuliaGeo/NetCDF.jl).

source

ClimateTools.ndgrid — Method.

X, Y = ndgrid(XV, YV)

This function creates a 2-D mesh-grid in a format consistent with Matlab's function ndgrid(). XV and YV are vectors.

source

ClimateTools.pr_timefactor — Method.

function pr_timefactor(rez::String)

Return the time factor that should be applied to precipitation to get accumulation for resolution "rez"

source

ClimateTools.prcp1 — Method.

prcp1(data::AbstractArray, timevector::StepRange{Date,Base.Dates.Day})

Annual number with preciptation >= 1 mm. This function returns a boolean vector. true if the data is higher or equal to 1 and false otherwise.

source

ClimateTools.prcp1 — Method.

prcp1(C::ClimGrid)

Annual number with preciptation >= 1 mm. This function returns a ClimGrid.

source

ClimateTools.qqmap — Method.

qqmap(obs::Array{N, 1} where N, ref::Array{N, 1} where N, fut::Array{N, 1} where N; method="Additive", detrend=true, window=15, rankn=50, thresnan=0.1, keep_original=false, interp::Function = Linear(), extrap::Function = Flat())

Quantile-Quantile mapping bias correction for single vector. This is a low level function used by qqmap(A::ClimGrid ..), but can work independently.

source

ClimateTools.qqmap — Method.

qqmap(obs::ClimGrid, ref::ClimGrid, fut::ClimGrid; method="Additive", detrend=true, window::Int=15, rankn::Int=50, thresnan::Float64=0.1, keep_original::Bool=false, interp::Function = Linear(), extrap::Function = Flat())

Quantile-Quantile mapping bias correction. For each julian day of the year (+/- window size), a transfer function is estimated through an empirical quantile-quantile mapping.

The quantile-quantile transfer function between ref and obs is etimated on a julian day (and grid-point) basis with a moving window around the julian day. Hence, for a given julian day, the transfer function is then applied to the fut dataset for a given julian day.

Options

method::String = "Additive" (default) or "Multiplicative". Additive is used for most climate variables. Multiplicative is usually bounded variables such as precipitation and humidity.

detrend::Bool = true (default). A 4th order polynomial is adjusted to the time series and the residuals are corrected with the quantile-quantile mapping.

window::Int = 15 (default). The size of the window used to extract the statistical characteristics around a given julian day.

rankn::Int = 50 (default). The number of bins used for the quantile estimations. The quantiles uses by default 50 bins between 0.01 and 0.99. The bahavior between the bins is controlled by the interp keyword argument. The behaviour of the quantile-quantile estimation outside the 0.01 and 0.99 range is controlled by the extrap keyword argument.

thresnan::Float64 = 0.1 (default). The fraction is missing values authorized for the estimation of the quantile-quantile mapping for a given julian days. If there is more than treshnan missing values, the output for this given julian days returns NaNs.

keep_original::Bool = false (default). If keep_original is set to true, the values are set to the original values in presence of too many NaNs.

interp = Interpolations.Linear() (default). When the data to be corrected lies between 2 quantile bins, the value of the transfer function is linearly interpolated between the 2 closest quantile estimation. The argument is from Interpolations.jl package.

extrap = Interpolations.Flat() (default). The bahavior of the quantile-quantile transfer function outside the 0.01-0.99 range. Setting it to Flat() ensures that there is no "inflation problem" with the bias correction. The argument is from Interpolation.jl package.

source

ClimateTools.shapefile_coords — Method.

shapefile_coords(poly::Shapefile.Polygon)

This function return the polygons contained in shp.shapes[i]. It returns the x and y coordinates vectors.

See also shapefile_coords_poly, which returns a polygon that ca be used for data extraction of the nc2julia.

source

ClimateTools.shapefile_coords_poly — Method.

shapefile_coords_poly(poly::Shapefile.Polygon)

Return the polygons contained in shp.shapes[i]. It returns an array containing the polygons.

See also shapefile_coords, which returns vectors as opposed to array. Returned polygon is consistent with the data extraction of the nc2julia function.

source

ClimateTools.spatialsubset — Method.

spatialsubset(C::ClimGrid, poly::Array{N, 2})

Returns the spatial subset of ClimGrid C. The spatial subset is defined by the polygon poly, defined on a -180, +180 longitude reference.

source

ClimateTools.sumleapyear — Method.

Number of leap years in date vector

sumleapyear(dates::StepRange{Date,Base.Dates.Day})

sumleapyear(initDate::Date, timeRaw)

source

ClimateTools.summerdays — Method.

summerdays(data::AbstractArray, time::StepRange{Date,Base.Dates.Day})

SD, Number of summer days: Annual count of days when TX (daily maximum temperature) > 25 degree Celsius.

Let TX[i,j] be daily maximum temperature on day i in year j. Count the number of days where:

TX[i,j] > 25 Celsius.

source

ClimateTools.summerdays — Method.

summerdays(C::ClimGrid)

SD, Number of summer days: Annual count of days when TX (daily maximum temperature) > 25 degree Celsius.

Let TX[i,j] be daily maximum temperature on day i in year j. Count the number of days where:

TX[i,j] > 25 Celsius.

source

ClimateTools.temporalsubset — Method.

function temporalsubset(C::ClimGrid, startdate::Date, enddate::Date)

Returns the temporal subset of ClimGrid C. The temporal subset is defined by a start and end date.

source

ClimateTools.timeindex — Method.

timeindex(timeVec, start_date, end_date, freq)

Return the index of time vector specified by start_date and end_date. Provide timestep "freq" to account for monthly timestep.

source

ClimateTools.timeresolution — Method.

timeresolution(timevec::Array{N,1} where N)

Return the time resolution of the vector timevec.

source

ClimateTools.tropicalnights — Method.

tropicalnights(data::AbstractArray, time::StepRange{Date,Base.Dates.Day})

TropicalNights, Number of tropical nights: Annual count of days when TN (daily maximum temperature) > 20 degree Celsius.

Let TN[i,j] be daily minimum temperature on day i in year j. Count the number of days where:

TN[i,j] > 20 Celsius.

source

ClimateTools.tropicalnights — Method.

tropicalnights(C::ClimGrid)

TropicalNights, Number of tropical nights: Annual count of days when TN (daily maximum temperature) > 20 degree Celsius.

Let TN[i,j] be daily minimum temperature on day i in year j. Count the number of days where:

TN[i,j] > 20 Celsius.

source

ClimateTools.windnr — Method.

windnr(p, poly::Matrix)

Determines the winding number of a point and a polygon, i.e. how many times a polygon winds around the point.

It follows Dan Sunday: http://geomalgorithms.com/a03-_inclusion.html.

source

PyPlot.plot — Method.

plot(C::ClimGrid)

Plots the spatial average timeserie of ClimGrid C.

source

ClimateTools.ClimGrid — Type.

ClimGrid{A <: AxisArray}

In-memory representation of Climate Forecast netCDF files.

source

ClimateTools.ClimGrid — Method.

ClimGrid(data; longrid=[], latgrid=[], msk=[], grid_mapping=Dict(), dimension_dict=Dict(), model="NA", frequency="NA", experiment="NA", run="NA", project="NA", institute="NA", filename="NA", dataunits="NA", latunits="NA", lonunits="NA", variable="NA", typeofvar="NA", typeofcal="NA", varattribs=Dict(), globalattribs=Dict())

Constructor of the ClimGrid function. Data is an AxisArray. Everything else is optional, but usually needed for further processing (mapping, interpolation, etc...).

source