Maps

Maps

Mapping the ClimGrid type

Mapping climate information can be done by using mapclimgrid.

ClimateTools.mapclimgrid — Function.

mapclimgrid(C::ClimGrid; region::String="auto", poly, level, mask, caxis, start_date::Date, end_date::Date)

Maps the time-mean average of ClimGrid C.

Optional keyworkd includes precribed regions (see list below), spatial clipping by polygon (keyword poly) or mask (keyword mask, an array of NaNs and 1.0 of the same dimension as the data in ClimGrid C), start_date and end_date.

Arguments

Regions

Europe

NorthAmerica

Canada

Quebec, QuebecNSP

Americas

World, WorldAz, WorldEck4

Greenwich

source

mapclimgrid(; region::String="auto", poly, level, mask, caxis, start_date::Date, end_date::Date)

Empty map generator, when called without a ClimGrid as the positional argument.

source

<p align="center"> <img src="https://cloud.githubusercontent.com/assets/3630311/23712122/e97bd322-03ef-11e7-93da-749c961c4070.png?raw=true" width="771" height="388" alt="Precipitation example"/> </p>

Note that the function plots the climatological mean of the provided ClimGrid. Multiple options are available for region: World, Canada, Quebec and the default auto which use the maximum and minimum of the lat-long coordinates inside the ClimGrid structure. The user can also provide a polygon(s) and the mapclimgrid function will clip the grid points outside the specified polygon. Another option is to provide a mask (with dimensions identical to the spatial dimension of the ClimGrid data) which contains NaN and 1.0 and the data inside the ClimGrid struct will be clipped with the mask. Other regions will be added in the future, as well as the option to send a custom region defined by a lat-lon box.